// src/models/Itil/taskModels/lastTask-model.ts
export interface lastTask {
    taskUID:                string;
    taskName:               string;
    taskDateCreate:         string;
    taskPriority:           string;
    taskInitiator:          string;
    taskOrganization:       string;
    taskNowExecutor:        string;
    taskDateCompilet:       string;
    taskCurrentStage:       string;
    taskService:            string;
    taskOrganizationClient: string;
    taskNumber:             string;   
}